forge fmt

+++ src/IAction.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.0;
 
 interface IAction {
-    function onSend(address from, address to, address matchId, uint quantity, bytes calldata data) external;
-} 
+    function onSend(address from, address to, address matchId, uint256 quantity, bytes calldata data) external;
+}
